+++ Sources/Shared/Extensions/Array+Extension.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+public extension Array where Element: Equatable {
+    func group<U: Hashable>(by transform: (Element) -> U) -> [U: [Element]] {
+        return reduce([:]) { dictionary, element in
+            var dictionary = dictionary
+            let key = transform(element)
+            dictionary[key] = (dictionary[key] ?? []) + [element]
+            return dictionary
+        }
+    }
+}